fixing
src\mapiclient\dump.c(414) : warning C4267: '=' : conversion from 'size_t' to 'int', possible loss of data

use ssize_t instead of int

+++ clients/src/mapiclient/dump.c
@@ -293,7 +293,7 @@ describe_table(Mapi mid, char *schema, char *tname, stream *toConsole, int forei
 	char *view = NULL;
 	size_t maxquerylen;
 	char *sname = NULL;
-	int cap;
+	ssize_t cap;
 #define CAP(X) ((cap = (X)) < 0 ? 0 : cap)
